+++ src/listbee/_client.py
@@ -9,7 +9,6 @@ from listbee.resources.customers import AsyncCustomers, Customers
 from listbee.resources.files import AsyncFiles, Files
 from listbee.resources.listings import AsyncListings, Listings
 from listbee.resources.orders import AsyncOrders, Orders
-from listbee.resources.store import AsyncStore, Store
 from listbee.resources.stripe import AsyncStripe, Stripe
 from listbee.resources.utility import AsyncUtility, Utility
 from listbee.resources.webhooks import AsyncWebhooks, Webhooks
@@ -46,7 +45,6 @@ class ListBee(SyncClient):
     def __init__(self, **kwargs: Any) -> None:
         super().__init__(**kwargs)
         self.bootstrap = Bootstrap(self)
-        self.store = Store(self)
         self.customers = Customers(self)
         self.files = Files(self)
         self.listings = Listings(self)
@@ -110,7 +108,6 @@ class AsyncListBee(AsyncClient):
     def __init__(self, **kwargs: Any) -> None:
         super().__init__(**kwargs)
         self.bootstrap = AsyncBootstrap(self)
-        self.store = AsyncStore(self)
         self.customers = AsyncCustomers(self)
         self.files = AsyncFiles(self)
         self.listings = AsyncListings(self)